feat: implement course permissions and authorization checks

+++ app/Http/Controllers/Api/Admin/CourseController.php
@@ -8,17 +8,18 @@ use App\Data\Course\CourseResponseData;
 use App\Http\Controllers\Controller;
 use App\Models\Course;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\QueryBuilder\QueryBuilder;
-use function Pest\Laravel\json;
 
 class CourseController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      */
     public function index():  ApiResponseInterface
     {
+        Gate::authorize('viewAny', Course::class);
         $courses = QueryBuilder::for(Course::class)
             ->allowedFilters(['slug', 'name', 'short_name', 'status'])
             ->allowedSorts(['slug', 'name', 'short_name', 'status'])
@@ -33,6 +34,8 @@ class CourseController extends Controller
      */
     public function store(CourseData $data):  ApiResponseInterface
     {
+        Gate::authorize('create',Course::class);
+
         $course = Course::query()->create($data->all());
 
         return response()->created(CourseResponseData::from($course)->toArray());
@@ -43,6 +46,8 @@ class CourseController extends Controller
      */
     public function show(Course $course):  ApiResponseInterface
     {
+        Gate::authorize('view',$course);
+
         return response()->success(CourseResponseData::from($course)->toArray());
     }
 
@@ -51,6 +56,8 @@ class CourseController extends Controller
      */
     public function edit(Course $course):  ApiResponseInterface
     {
+        Gate::authorize('update', $course);
+
         return response()->success(CourseData::from($course)->toArray());
     }
 
@@ -59,6 +66,7 @@ class CourseController extends Controller
      */
     public function update(CourseData $data, Course $course):  ApiResponseInterface
     {
+        Gate::authorize('update', $course);
         $course->update($data->all());
         return response()->success(CourseResponseData::from($course)->toArray());
     }
@@ -68,6 +76,7 @@ class CourseController extends Controller
      */
     public function destroy(Course $course): JsonResponse
     {
+        Gate::authorize('delete', $course);
         $course->delete();
         return response()->noContentJson();
     }
